+++ packages/common/base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/common-sdk-base
 
+## 0.5.1
+
+### Patch Changes
+
+- 8e4d900: Add support for world-chain
+
+## 0.5.0
+
+### Minor Changes
+
+- e597884: Remove support for Cardano
+
 ## 0.4.2
 
 ### Patch Changes