+++ packages/server/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @crossmint/server-sdk
 
+## 1.2.22
+
+### Patch Changes
+
+- Updated dependencies [8e4d900]
+  - @crossmint/common-sdk-base@0.5.1
+  - @crossmint/common-sdk-auth@1.0.32
+
+## 1.2.21
+
+### Patch Changes
+
+- Updated dependencies [e597884]
+  - @crossmint/common-sdk-base@0.5.0
+  - @crossmint/common-sdk-auth@1.0.31
+
 ## 1.2.20
 
 ### Patch Changes